refactor: organize graph controll logic
